壁に近づきすぎ問題を解決
