+++ read_map.c
@@ -6,7 +6,7 @@
 /*   By: yookamot <yookamot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/19 15:13:58 by yookamot          #+#    #+#             */
-/*   Updated: 2025/08/21 23:03:12 by yookamot         ###   ########.fr       */
+/*   Updated: 2025/08/30 20:38:14 by yookamot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	read_map(t_data *data)
 		exit(1);
 	data->map[0] = strdup("11111");
 	data->map[1] = strdup("10001");
-	data->map[2] = strdup("10201");
+	data->map[2] = strdup("10101");
 	data->map[3] = strdup("10001");
 	data->map[4] = strdup("10N01");
 	data->map[5] = strdup("10001");
